+++ src/index.ts
@@ -1,29 +1,17 @@
-export type SyncCheckFn<T> = T extends Promise<any> ? never : (iteration: number) => T;
-export type ASyncCheckFn<T> = (iteration: number) => Promise<T>;
+export type SyncCheckFn<T> = T extends Promise<any> ? never : (iteration: number, delay: number) => T;
+export type ASyncCheckFn<T> = (iteration: number, delay: number) => Promise<T>;
 export type CheckFn<T> = ASyncCheckFn<T> | SyncCheckFn<T>;
 
 export type Jitter = 'none' | 'full';
 
-export interface IBusyWaitExponentialBackoffOptions {
-    initialSleepTime: number;
-    multiplier?: number;
-    maxDelay?: number;
-    jitter?: Jitter;
-    type: 'EXPONENT';
-}
-
-export interface IBusyWaitLinearBackoffOptions {
-    sleepTime: number;
-    type: 'LINEAR';
-}
-
-export type IBusyWaitBackoffOptions = IBusyWaitLinearBackoffOptions | IBusyWaitExponentialBackoffOptions;
-
 export interface IBusyWaitOptions {
+    sleepTime: number;
     failMsg?: string;
-    waitFirst?: boolean;
     maxChecks?: number;
-    backoff: IBusyWaitBackoffOptions;
+    waitFirst?: boolean;
+    multiplier?: number;
+    jitter?: Jitter;
+    maxDelay?: number;
 }
 
 export interface IBusyWaitResult<T> {
@@ -34,34 +22,28 @@ export interface IBusyWaitResult<T> {
     result: T;
 }
 
-const isLinearBackoffOptions = (x: IBusyWaitBackoffOptions): x is IBusyWaitLinearBackoffOptions => {
-    return x.type === 'LINEAR';
-}
+type Resolve<T> = (result: T) => void;
+type Reject = (error: Error) => void;
+type Delayer = (iteration: number) => number;
 
 const isFunction = (value: any): value is () => any => toString.call(value) === '[object Function]';
 const isNumber = (value: any): value is number => typeof value === 'number';
 
-const getOptions = <T>(checkFn: CheckFn<T>, _options: IBusyWaitOptions): IBusyWaitOptions => {
+const getAndValidateOptions = <T>(checkFn: CheckFn<T>, _options: IBusyWaitOptions): IBusyWaitOptions => {
     const options = Object.assign({}, _options);
     if (isNumber(options.maxChecks) && (isNaN(options.maxChecks) || options.maxChecks < 1)) {
-        throw new Error('maxChecks must be a valid integer greater than 0');
+        throw new Error('maxChecks must be a valid integer greater than 1');
     }
-    if (isLinearBackoffOptions(options.backoff)) {
-        if (isNaN(options.backoff.sleepTime) || options.backoff.sleepTime < 1) {
-            throw new Error('sleepTime must be a valid integer greater than 0');
-        }
-    } else {
-        if (isNaN(options.backoff.initialSleepTime) || options.backoff.initialSleepTime < 1) {
-            throw new Error('initialSleepTime must be a valid integer greater than 0');
-        }
-        if (isNumber(options.backoff.multiplier) &&
-            (isNaN(options.backoff.multiplier) || options.backoff.multiplier < 1)) {
-            throw new Error('multiplier must be a valid integer greater than 0');
-        }
-        if (isNumber(options.backoff.maxDelay) &&
-            (isNaN(options.backoff.maxDelay) || options.backoff.maxDelay < 1)) {
-            throw new Error('maxDelay must be a valid integer greater than 0');
-        }
+    if (isNaN(options.sleepTime) || options.sleepTime < 1) {
+        throw new Error('sleepTime must be a valid integer greater than 1');
+    }
+    if (isNumber(options.multiplier) &&
+        (isNaN(options.multiplier) || options.multiplier < 1)) {
+        throw new Error('multiplier must be a valid integer greater than 1');
+    }
+    if (isNumber(options.maxDelay) &&
+        (isNaN(options.maxDelay) || options.maxDelay < 1)) {
+        throw new Error('maxDelay must be a valid integer greater than 1');
     }
     if (!checkFn || !isFunction(checkFn)) {
         throw new Error('checkFn must be a function');
@@ -69,51 +51,56 @@ const getOptions = <T>(checkFn: CheckFn<T>, _options: IBusyWaitOptions): IBusyWa
     return options;
 };
 
-const wrapSyncMethod = <T>(checkFn: CheckFn<T>): ASyncCheckFn<T> => async (iteration: number) => checkFn(iteration);
+const wrapSyncMethod = <T>(checkFn: CheckFn<T>): ASyncCheckFn<T> =>
+    async (iteration: number, delay: number) => checkFn(iteration, delay);
 
-const eventLoop = <T>(checkFn: ASyncCheckFn<T>, options: IBusyWaitOptions, delayer: Delayer)
-    : Promise<IBusyWaitResult<T>> => {
-    return new Promise((resolve, reject) => {
-        let iteration = 0;
-        const start = Date.now();
-        const iterationCheck = async () => {
-            iteration++;
-            try {
-                const result = await checkFn(iteration);
-                return resolve({
-                    backoff: {
-                        iterations: iteration,
-                        time: Date.now() - start,
-                    },
-                    result,
-                });
-            } catch (e) {
-                if (options.maxChecks && iteration === options.maxChecks) {
-                    return reject(new Error(options.failMsg || 'Exceeded number of iterations to wait'));
-                }
-                setTimeout(iterationCheck, delayer(iteration));
-            }
-        };
-        setTimeout(iterationCheck, options.waitFirst ? delayer(iteration) : 0);
+const unwrapPromise = <T>() => {
+    let resolve: Resolve<T>;
+    let reject: Reject;
+    const promise = new Promise<T>((_resolve, _reject) => {
+        resolve = _resolve;
+        reject = _reject;
     });
-};
+    // @ts-expect-error
+    return {promise, resolve, reject};
+}
 
-export const busywait = async <T>(checkFn: CheckFn<T>, _options: IBusyWaitOptions): Promise<IBusyWaitResult<T>> => {
-    const options = getOptions(checkFn, _options);
-    const delayer = getDelayer(options.backoff);
-    return eventLoop(wrapSyncMethod(checkFn), options, delayer);
-};
-
-type Delayer = (iteration: number) => number;
-
-const getDelayer = (options: IBusyWaitBackoffOptions): Delayer => (iteration: number) => {
-    if (isLinearBackoffOptions(options)) {
-        return options.sleepTime;
-    }
-    let delay = options.initialSleepTime * Math.pow(options.multiplier || 2, iteration);
+const getDelayer = (options: IBusyWaitOptions): Delayer => (iteration: number) => {
+    let delay = options.sleepTime * Math.pow(options.multiplier || 1, iteration);
     delay = Math.min(delay, options.maxDelay || Infinity);
     if (options.jitter === 'full') {
         return Math.round(Math.random() * delay);
     }
     return delay;
 }
+
+export const busywait = async <T>(_checkFn: CheckFn<T>, _options: IBusyWaitOptions): Promise<IBusyWaitResult<T>> => {
+    const options = getAndValidateOptions(_checkFn, _options);
+    const delayer = getDelayer(options);
+    const checkFn = wrapSyncMethod(_checkFn);
+    let iteration = 0;
+    const start = Date.now();
+    let delay = options.waitFirst ? delayer(iteration) : 0;
+    const {promise, resolve, reject} = unwrapPromise<IBusyWaitResult<T>>();
+    const iterationCheck = async () => {
+        iteration++;
+        try {
+            const result = await checkFn(iteration, delay);
+            return resolve({
+                backoff: {
+                    iterations: iteration,
+                    time: Date.now() - start,
+                },
+                result,
+            });
+        } catch (e) {
+            if (options.maxChecks && iteration === options.maxChecks) {
+                return reject(new Error(options.failMsg || 'Exceeded number of iterations to wait'));
+            }
+            delay = delayer(iteration);
+            setTimeout(iterationCheck, delay);
+        }
+    };
+    setTimeout(iterationCheck, delay);
+    return promise;
+};